+++ docs/Changelog.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## 2.0.15
+- Added temporary solution for problem with accepting invitations (in few cases). Thanks [Terbau](https://github.com/Terbau) for [PR](https://github.com/SzymonLisowiec/node-epicgames-client/pull/52).
+- Added [`waitForEvent()`](/Communicator.html#waitforevent-event-timeout) method to `Communicator`. Thanks [Terbau](https://github.com/Terbau) for [PR](https://github.com/SzymonLisowiec/node-epicgames-client/pull/52).
+- Added `getFriendStatus(id)` method for `Launcher` (`Client`). Thanks [Terbau](https://github.com/Terbau) for [PR](https://github.com/SzymonLisowiec/node-epicgames-client/pull/52).
+- Added Nintendo Switch to platform's enumerator (`EPlatform`).
+- Added communicator's events:
+  - `friend#ID:request`
+  - `friend#ID:added`
+  - `friend#ID:removed`
+  - `friend#ID:status`
+  - `friend#ID:message`
+- Updated launcher's `netCL`, `build` and `engineBuild`.
+
 ## 2.0.14
 - Updated `enableTwoFactor()` and `disableTwoFactor()` methods.